+++ data/src/main/java/com/ahuaman/data/mapper/PokemonDetailMapper.kt
@@ -9,9 +9,13 @@ fun PokemonDetailResponse.toDomain(): PokemonDetailDomainModel {
         id = id,
         name = name,
         imageUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/$id.png",
-        weightKg = weight / 10.0, // Logic stays in Data Layer
-        heightM = height / 10.0,  // Logic stays in Data Layer
-        types = types.map { it.type.name },
-        stats = stats.map { PokemonStat(name = it.stat.name, value = it.baseStat) }
+        weightKg = weight / 10.0,
+        heightM = height / 10.0,
+        // Senior Move: Sanitize the data at the boundary
+        types = types?.mapNotNull { it.type?.name }.orEmpty(),
+        stats = stats?.mapNotNull {
+            val statName = it.stat?.name ?: "unknown"
+            PokemonStat(name = statName, value = it.baseStat)
+        }.orEmpty()
     )
 }