+++ data/src/main/java/com/ahuaman/data/models/pokemon/PokemonResponse.kt
@@ -4,11 +4,11 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PokemonResponse(
-    val results: List<PokemonDataModel>
+    val results: List<PokemonDataModel>? = null // Nullable for safety
 )
 
 @Serializable
 data class PokemonDataModel(
-    val name: String,
-    val url: String
+    val name: String? = null, // Even name/url can be missing in corrupt data
+    val url: String? = null
 )