+++ testing/src/main/java/com/ahuaman/testing/pokemons/PokemonMocks.kt
@@ -1,6 +1,8 @@
 package com.ahuaman.testing.pokemons
 
+import com.ahuaman.domain.model.PokemonDetailDomainModel
 import com.ahuaman.domain.model.PokemonDomainModel
+import com.ahuaman.domain.model.PokemonStat
 
 object PokemonMocks {
 
@@ -71,5 +73,19 @@ object PokemonMocks {
             name = "Dragonite",
             imageUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/149.png"
         ),
+    )
+
+    val bulbasaurDetail = PokemonDetailDomainModel(
+        id = 1,
+        name = "bulbasaur",
+        imageUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/1.png",
+        weightKg = 6.9,
+        heightM = 0.7,
+        types = listOf("grass", "poison"),
+        stats = listOf(
+            PokemonStat("hp", 45),
+            PokemonStat("attack", 49),
+            PokemonStat("defense", 49)
+        )
     )
 }